Initial upload of files
Very basic functionality as of now. CSS will be added in future, the main purpose of this so far is to experiment with JS. I plan to add buttons for what the skill being tested is, so you could just select attack, damage, bluff, etc. Additionally, I will change the skill level entries to number drop downs to avoid issues with entering something other than a number.